+++ test/src/json.test.ts
@@ -14,17 +14,6 @@ describe("createJsonResponse test", (): void => {
         });
     });
 
-    it("should work with failure", async (): Promise<void> => {
-        const res: Response = createJsonResponse({
-            success: false,
-            error: {
-                code: "server",
-            },
-        });
-
-        expect(res.status).toBe(400);
-    });
-
     it("should work with data", async (): Promise<void> => {
         const res: Response = createJsonResponse({
             data: {
@@ -44,10 +33,55 @@ describe("createJsonResponse test", (): void => {
         });
     });
 
+    it("should work as failure", async (): Promise<void> => {
+        const res: Response = createJsonResponse({
+            error: {
+                code: "server",
+            },
+        });
+
+        expect(res.status).toBe(400);
+
+        const body: string = await res.text();
+
+        expect(JSON.parse(body)).toStrictEqual({
+            success: false,
+            error: {
+                code: "server",
+            },
+        });
+    });
+
+    it("should work as failure with success: false", async (): Promise<void> => {
+        const res: Response = createJsonResponse({
+            success: false,
+            error: {
+                code: "server",
+            },
+        });
+
+        expect(res.status).toBe(400);
+
+        const body: string = await res.text();
+
+        expect(JSON.parse(body)).toStrictEqual({
+            success: false,
+            error: {
+                code: "server",
+            },
+        });
+    });
+
     it("should work with customization", async (): Promise<void> => {
         const res: Response = createJsonResponse({
             success: false,
             status: 500,
+            headers: [
+                [
+                    "X-Test",
+                    "test",
+                ],
+            ],
             error: {
                 code: "server",
                 field: "server",
@@ -56,6 +90,8 @@ describe("createJsonResponse test", (): void => {
         });
 
         expect(res.status).toBe(500);
+
+        expect(res.headers.get("X-Test")).toBe("test");
 
         const body: string = await res.text();